+++ pccommon/middleware.py
@@ -1,0 +1,35 @@
+import logging
+from typing import Callable, Awaitable
+
+from fastapi import HTTPException, Request, Response
+from pccommon.logging import request_to_path
+from pccommon.tracing import (
+    HTTP_URL,
+    HTTP_METHOD,
+    HTTP_PATH
+)
+
+logger = logging.getLogger(__name__)
+
+async def handle_exceptions(
+    request: Request, call_next: Callable[[Request], Awaitable[Response]]
+) -> Response:
+    try:
+        raise Exception("ha")
+        return call_next(request)
+    except HTTPException:
+        raise
+    except Exception as e:
+        logger.exception(
+            "Exception when handling request",
+            extra={
+                "custom_dimensions": {
+                    "stackTrace": f"{e}",
+                    HTTP_URL: str(request.url),
+                    HTTP_METHOD: str(request.method),
+                    HTTP_PATH: request_to_path(request),
+                    "service": "tiler",
+                }
+            },
+        )
+        raise
