+++ pctiler/pctiler/middleware.py
@@ -133,27 +133,3 @@ async def count_data_requests(request: Request, call_next):  # type: ignore
             request,
         )
     return await call_next(request)
-
-
-async def handle_exceptions(
-    request: Request, call_next: Callable[[Request], Awaitable[Response]]
-) -> Response:
-    try:
-        raise Exception("ha")
-        return call_next(request)
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.exception(
-            "Exception when handling request",
-            extra={
-                "custom_dimensions": {
-                    "stackTrace": f"{e}",
-                    HTTP_URL: str(request.url),
-                    HTTP_METHOD: str(request.method),
-                    HTTP_PATH: request_to_path(request),
-                    "service": "tiler",
-                }
-            },
-        )
-        raise
